+++ tools/purge.js
@@ -40,9 +40,13 @@ const cdnConfig = {
     }
 };
 
-const branch = process.env.TRAVIS_BRANCH;
+let branch = process.env.TRAVIS_BRANCH;
 if(!branch) {
     throw new Error('Invalid branch: ' + branch);
+}
+
+if(branch === 'master') {
+    branch = 'development';
 }
 
 const commit = process.env.TRAVIS_COMMIT;
@@ -55,9 +59,8 @@ const paths = [
     '/test/config.json'
 ];
 
-const urlRoot = '/webview/' + branch;
-const urlsFromPath = (base_urls, path, addHttps) => {
-    return base_urls.map((baseUrl) => {
+const urlsFromPath = (urlRoot, baseUrls, path, addHttps) => {
+    return baseUrls.map((baseUrl) => {
         return (addHttps ? 'https://' : '') + baseUrl + urlRoot + path;
     });
 };
@@ -86,9 +89,9 @@ const checkConfigJson = (url, version) => {
     return doFetch();
 };
 
-let purgeAkamai = () => {
-    let urls = flatten(paths.map(function(path) {
-        return urlsFromPath(cdnConfig.akamai.base_urls, path, true);
+let purgeAkamai = (urlRoot) => {
+    let urls = flatten(paths.map(path => {
+        return urlsFromPath(urlRoot, cdnConfig.akamai.base_urls, path, true);
     }));
 
     console.log('Starting Akamai purge of: ');
@@ -120,9 +123,9 @@ let purgeAkamai = () => {
     });
 };
 
-let purgeHighwinds = () => {
-    let urls = flatten(paths.map(function(path) {
-        return urlsFromPath(cdnConfig.highwinds.base_urls, path, true);
+let purgeHighwinds = (urlRoot) => {
+    let urls = flatten(paths.map(path => {
+        return urlsFromPath(urlRoot, cdnConfig.highwinds.base_urls, path, true);
     }));
 
     console.log('Starting Highwinds purge of: ');
@@ -155,9 +158,9 @@ let purgeHighwinds = () => {
     });
 };
 
-let purgeChinaNetCenter = () => {
-    let urls = flatten(paths.map(function(path) {
-        return urlsFromPath(cdnConfig.chinanetcenter.base_urls, path, false);
+let purgeChinaNetCenter = (urlRoot) => {
+    let urls = flatten(paths.map(path => {
+        return urlsFromPath(urlRoot, cdnConfig.chinanetcenter.base_urls, path, false);
     }));
 
     console.log('Starting ChinaNetCenter purge of: ');
@@ -187,10 +190,15 @@ let purgeChinaNetCenter = () => {
     });
 };
 
+let urlRoot = '/webview/' + branch;
+if(branch === '2.0.6-ios') {
+    urlRoot = '/webview/master';
+}
+
 Promise.all([
-    purgeAkamai(),
-    purgeHighwinds(),
-    purgeChinaNetCenter()
+    purgeAkamai(urlRoot),
+    purgeHighwinds(urlRoot),
+    purgeChinaNetCenter(urlRoot)
 ]).then(() => {
     console.log('Successfully purged all CDNs!');
 });